refact: rename apiPayload package to payload
- 기존 패키지명 `apiPayload` → `payload` 로 변경하여 가독성 개선
- 패키지 이름을 간결하게 리네이밍함으로써 코드 구조 정리
- 관련된 클래스 경로 모두 업데이트하여 일관성 유지
