--- src/main/kotlin/com/dobby/backend/presentation/api/dto/apiPayload/code/BaseCode.kt
@@ -1,4 +1,0 @@
-package com.dobby.backend.presentation.api.dto.apiPayload.code
-
-interface BaseCode {
-}
