+++ src/main/kotlin/com/dobby/backend/presentation/api/dto/payload/code/BaseCode.kt
@@ -1,0 +1,4 @@
+package com.dobby.backend.presentation.api.dto.payload.code
+
+interface BaseCode {
+}
