+++ src/main/kotlin/com/dobby/backend/presentation/api/dto/payload/code/ReasonDto.kt
@@ -1,4 +1,4 @@
-package com.dobby.backend.presentation.api.dto.apiPayload.code
+package com.dobby.backend.presentation.api.dto.payload.code
 
 import org.springframework.http.HttpStatus
